wip: compiler errors | update lambdas errs
