+++ src/components/transactions/history/List.styles.ts
@@ -1,3 +1,4 @@
+import { m } from 'motion/react';
 import styled from 'styled-components';
 
 export const ListWrapper = styled.div`
@@ -25,4 +26,26 @@ export const FilterWrapper = styled.div`
     display: flex;
     gap: 12px;
     flex-shrink: 0;
+`;
+
+export const LoadingText = styled(m.div)`
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    font-size: 12px;
+    font-weight: 500;
+    text-align: center;
+    z-index: 1;
+
+    margin-top: 7px;
+
+    background-color: ${({ theme }) => (theme.mode === 'dark' ? '#2E2E2E' : '#fff')};
+    color: ${({ theme }) => (theme.mode === 'dark' ? '#fff' : '#000')};
+    padding: 8px 20px;
+    border-radius: 100px;
+    width: max-content;
+    max-width: 220px;
+
+    box-shadow: 0 0 115px 0 rgba(0, 0, 0, 0.35);
 `;
